test: move async test to schema test
